Use tmix rather than tblend.
Using tblend was less efficient. It would blend frames 0+1, 1+2, 2+3,
and so on. After that it would drop every second frame. So half of the
work for blending frames would have been thrown out! The tmix approach
is better, because it doens't do work that is later thrown out.
Meanwhile, the effect is the same.